requirements.txt
google-cloud-bigquery
google-auth
python-dotenv
pandas
scikit-learn
matplotlib
seaborn
openai
db-dtypes
